refactor(hardware-profile): split platform discovery
